tqmls1028a: implement boot source detection
Implement boot source detection for SD/eMMC boot. SPI-NOR is left out, as
a separate BL2 will be needed anyways (if we want avoid to patch the ATF
more heavily than necessary).

Byte offsets in the PBI are added as comments to make the jump
instructions and total length a bit clearer. This is implemented for
SPI-NOR as well.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>