tqmls1028a: default: rcw_common: use IIC2 pins as gpios
When the IIC2_PMUX pins are used for SDHC1 these pins are used as
write-protect and card-detect. But we have the write-protect pin
connected to the interrupt output of the port expanders. So when all
interrupts where cleared by reading the inputs of the port expanders the
sd-card interface was turned to write-protected and caused timeouts when
writing.

Fix this issue by configuring the two pins as gpio.

Signed-off-by: Gregor Herburger <gregor.herburger@tq-group.com>