tqmls1028a: add Secure-Boot-enabled versions of all RCW variants
This adds a Secure Boot version for each RCW variant, with the "_sben"
suffix expected by the qoriq-atf Yocto recipe.

SB_EN is set in the RCW to enable Secure Boot support in development
mode. As long as Secure Boot is not enabled unconditionally by programming
the ITS fuse, the LS1028A will accept both secured and unsecured images,
and will continue boot with an error message when verification of a
secured image fails.

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>