tqmlx2160a: mux EC1/EC2 (RGMII) pins as GPIO when SERDES2 is SGMII
EC1 and EC2 correspond to DPMAC 17 and 18, but when SERDES2 is
configured for SGMII (configurations 7 and 11 in our RCWs), the same MACs
are also used for SGMII. We have been informed by NXP that in these
configurations, using DPMAC 17 and 18 for RGMII will cause errors in
certain DPAA2 setups.

Disable the RGMII pins in these configurations by muxing them as GPIOs.
The total number of available GBit Ethernet ports in unchanged, but the
physical ports corresponding to DPMAP 17 and 18 change (and
phy-mode/phy-handle in U-Boot/Linux Device Trees need to be updated).

Signed-off-by: Matthias Schiffer <matthias.schiffer@tq-group.com>